Added feature from Si Chen to make it possible to select whether you want the files deleted or not in an import dir; will probably change this to default to checked as not deleting files makes it harder to handle error conditions in large sets of files (ie best to make sure you have a copy somewhere and then run this), but this will be nice to make it a little more clear what the thing will do...; Jira #OFBIZ-317
git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@5253 deaff617-5be1-0310-a1b7-ec6878f92ba9

+++ framework/webtools/webapp/webtools/entity/xmldsimportdir.jsp
@@ -38,6 +38,7 @@
   boolean mostlyInserts = request.getParameter("mostlyInserts") != null;
   boolean keepStamps = request.getParameter("maintainTimeStamps") != null;
   boolean createDummyFks = request.getParameter("createDummyFks") != null;
+  boolean deleteFiles = request.getParameter("deleteFiles") != null;
 
   String txTimeoutStr = UtilFormatOut.checkEmpty(request.getParameter("txTimeout"), "7200");
   String filePauseStr = UtilFormatOut.checkEmpty(request.getParameter("filePause"), "0");
@@ -71,6 +72,7 @@
     Mostly Inserts?:<INPUT type="checkbox" name='mostlyInserts' <%=mostlyInserts?"checked":""%>>
     Maintain Timestamps?:<INPUT type="checkbox" name='maintainTimeStamps' <%=keepStamps?"checked":""%>>
     Create "Dummy" FKs?:<INPUT type="checkbox" name='createDummyFks' <%=createDummyFks?"checked":""%>>
+    Delete Files Afterwards?:<INPUT type="checkbox" name='deleteFiles' <%=deleteFiles?"checked":""%>>
     </div>
     <div>TX Timeout Seconds:<INPUT type="text" size="6" value="<%=txTimeoutStr%>" name='txTimeout'>
     Pause (secs) between files:<INPUT type="text" size="6" value="<%=filePauseStr%>" name="filePause"></div>
@@ -127,7 +129,9 @@
 		        long numberRead = reader.parse(url);
 		        %><div>Got <%=numberRead%> entities from <%=curFile%></div><%
                 importedOne = true;
-                curFile.delete();
+                if (deleteFiles) {
+                    curFile.delete();    
+                }
 	        } catch (Exception ex){
                 %> <div>Error trying to read from <%=curFile%>: <%=ex%> <%
                 if (ex.toString().indexOf("referential integrity violation") > -1 ||
